Relocated assignment page + components to the correct directories
