+++ src/components/assignment/AssignmentDetails.tsx
