+++ src/components/assignment/ExpandedAssignmentView.css
