+++ src/components/assignment/ExpandedAssignmentView.tsx
